requirements.txt
<generator object iter_importers at 0x00000000029FD558>
########### __main__ ---%- <type 'zipimport.zipimporter'>
alabaster==0.7.11
amqp==2.3.2
aniso8601==3.0.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.1
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.4
bleach==2.1.3
cefpython3==57.0
celery==4.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
colorama==0.3.9
conda==4.5.11
configparser==3.5.0
cryptography==2.2.2
cycler==0.10.0
Cython==0.29
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.15
django-appconf==1.0.2
django-bootstrap3-datetimepicker==2.3
django-braces==1.8.1
django-celery-results==1.0.1
django-codemirror-widget==0.4.1
django-crispy-forms==1.6.0
django-embed-video==1.2.0
django-modelcluster==3.1
django-npm==1.0.0
django-registration-redux==1.4
Django-Select2==5.11.1
django-storages==1.7.1
django-taggit==0.22.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
django-yamlfield==1.0.3
djangorestframework==3.6.4
djangosaml2==0.17.2
dnspython==1.15.0
docker-py==1.10.6
docker-pycreds==0.3.0
docutils==0.14
draftjs-exporter==2.1.5
dulwich==0.19.6
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
graphviz==0.9
greenlet==0.4.15
h5py==2.8.0
html5lib==0.999999999
httpie==0.9.9
idna==2.6
imagesize==1.0.0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonfield==1.0.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.4.7.1
Keras==2.1.6
keras-applications==1.0.2
keras-preprocessing==1.0.1
Kivy==1.11.0.dev0
Kivy-examples==1.10.1
Kivy-Garden==0.1.4
kivy.deps.angle==0.1.6
kivy.deps.glew==0.1.10
kivy.deps.glew-dev==0.1.10
kivy.deps.gstreamer==0.1.13
kivy.deps.gstreamer-dev==0.1.13
kivy.deps.sdl2==0.1.18
kivy.deps.sdl2-dev==0.1.18
kiwisolver==1.0.1
kombu==4.1.0
lazy-object-proxy==1.3.1
linecache2==1.0.0
lxml==4.2.5
mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
monotonic==1.5
more-itertools==4.3.0
mpmath==1.0.0
mysql-connector-python==8.0.12
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.3
nose==1.3.7
notebook==5.7.0
numpy==1.12.0
numpydoc==0.8.0
olefile==0.46
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.2.1
pathlib2==2.3.2
pickleshare==0.7.4
Pillow==4.4.0
pkgconfig==1.4.0
plotly==3.3.0
pluggy==0.7.1
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.6
psycopg2==2.7.5
py==1.6.0
pyasn1==0.4.4
pycairo==1.17.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pydot==1.2.2
pyflakes==2.0.0
pygame==1.9.4
Pygments==2.2.0
pygpu==0.6.9
pylint==1.9.2
pymbolic==2018.1
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypiwin32==223
pyreadline==2.1
pysaml2==4.6.1
PySocks==1.6.8
pytest==3.8.2
python-dateutil==2.7.3
python-decouple==3.1
python-igraph==0.7.1.post6
python-memcached==1.59
python-slugify==1.2.5
pytools==2018.5.2
pytz==2017.3
pywin32==223
pywinpty==0.5.4
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
requests==2.18.4
retrying==1.3.3
rope==0.10.7
ruamel-yaml==0.15.37
scandir==1.7
scikit-learn==0.19.2
scipy==0.16.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
sleekxmpp==1.3.3
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
spyder==3.2.8
sympy==1.3
terminado==0.8.1
testpath==0.3.1
Theano==0.9.0
tornado==5.0.2
tqdm==4.25.0
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
Unidecode==1.0.22
unittest2==1.1.0
urllib3==1.22
vine==1.1.4
virtualenv==16.0.0
wagtail==1.13.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
Willow==1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
